refactor: structural improvements
