+++ lib/src/core/tensor_buffer.dart
@@ -5,6 +5,8 @@ import 'dtype.dart';
 import 'memory_format.dart';
 import 'tensor_storage.dart';
 
+part 'tensor_buffer_factory.dart';
+
 /// A multi-dimensional array view over typed data with shape and stride metadata.
 ///
 /// [TensorBuffer] provides a NumPy-like interface for tensor operations. It
@@ -347,56 +349,15 @@ class TensorBuffer {
     List<int> shape, {
     DType dtype = DType.float32,
     MemoryFormat memoryFormat = MemoryFormat.contiguous,
-  }) {
-    _validateShapeStatic(shape);
-    final numel = shape.fold(1, (a, b) => a * b);
-    final data = dtype.createBuffer(numel);
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable(shape),
-      memoryFormat: memoryFormat,
-    );
-  }
+  }) =>
+      _zerosImpl(shape, dtype: dtype, memoryFormat: memoryFormat);
 
   /// Creates a tensor filled with ones.
   static TensorBuffer ones(
     List<int> shape, {
     DType dtype = DType.float32,
-  }) {
-    _validateShapeStatic(shape);
-    final numel = shape.fold(1, (a, b) => a * b);
-    final data = dtype.createBuffer(numel);
-
-    for (int i = 0; i < numel; i++) {
-      switch (data) {
-        case final Float32List list:
-          list[i] = 1.0;
-        case final Float64List list:
-          list[i] = 1.0;
-        case final Int8List list:
-          list[i] = 1;
-        case final Int16List list:
-          list[i] = 1;
-        case final Int32List list:
-          list[i] = 1;
-        case final Int64List list:
-          list[i] = 1;
-        case final Uint8List list:
-          list[i] = 1;
-        case final Uint16List list:
-          list[i] = 1;
-        case final Uint32List list:
-          list[i] = 1;
-        case final Uint64List list:
-          list[i] = 1;
-      }
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  }) =>
+      _onesImpl(shape, dtype: dtype);
 
   /// Creates a tensor filled with a specific value.
   ///
@@ -409,41 +370,8 @@ class TensorBuffer {
     List<int> shape, {
     required double fillValue,
     DType dtype = DType.float32,
-  }) {
-    _validateShapeStatic(shape);
-    final numel = shape.fold(1, (a, b) => a * b);
-    final data = dtype.createBuffer(numel);
-
-    for (int i = 0; i < numel; i++) {
-      switch (data) {
-        case final Float32List list:
-          list[i] = fillValue;
-        case final Float64List list:
-          list[i] = fillValue;
-        case final Int8List list:
-          list[i] = fillValue.toInt();
-        case final Int16List list:
-          list[i] = fillValue.toInt();
-        case final Int32List list:
-          list[i] = fillValue.toInt();
-        case final Int64List list:
-          list[i] = fillValue.toInt();
-        case final Uint8List list:
-          list[i] = fillValue.toInt().clamp(0, 255);
-        case final Uint16List list:
-          list[i] = fillValue.toInt().clamp(0, 65535);
-        case final Uint32List list:
-          list[i] = fillValue.toInt();
-        case final Uint64List list:
-          list[i] = fillValue.toInt();
-      }
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  }) =>
+      _fullImpl(shape, fillValue: fillValue, dtype: dtype);
 
   /// Creates a tensor with random values uniformly distributed in [0, 1).
   ///
@@ -457,21 +385,8 @@ class TensorBuffer {
     List<int> shape, {
     DType dtype = DType.float32,
     int? seed,
-  }) {
-    _validateShapeStatic(shape);
-    final numel = shape.fold(1, (a, b) => a * b);
-    final data = Float32List(numel);
-    final rng = seed != null ? _SeededRandom(seed) : _SeededRandom.system();
-
-    for (int i = 0; i < numel; i++) {
-      data[i] = rng.nextDouble();
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  }) =>
+      _randomImpl(shape, dtype: dtype, seed: seed);
 
   /// Creates a tensor with random values from a standard normal distribution N(0, 1).
   ///
@@ -485,31 +400,8 @@ class TensorBuffer {
     List<int> shape, {
     DType dtype = DType.float32,
     int? seed,
-  }) {
-    _validateShapeStatic(shape);
-    final numel = shape.fold(1, (a, b) => a * b);
-    final data = Float32List(numel);
-    final rng = seed != null ? _SeededRandom(seed) : _SeededRandom.system();
-
-    // Box-Muller transform for normal distribution
-    for (int i = 0; i < numel; i += 2) {
-      final u1 = rng.nextDouble();
-      final u2 = rng.nextDouble();
-      // Avoid log(0)
-      final safeU1 = u1 < 1e-10 ? 1e-10 : u1;
-      final r = _sqrt(-2.0 * _log(safeU1));
-      final theta = 2.0 * _pi * u2;
-      data[i] = r * _cos(theta);
-      if (i + 1 < numel) {
-        data[i + 1] = r * _sin(theta);
-      }
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  }) =>
+      _randnImpl(shape, dtype: dtype, seed: seed);
 
   /// Creates a 2D identity matrix.
   ///
@@ -523,26 +415,8 @@ class TensorBuffer {
     int n, {
     int? m,
     DType dtype = DType.float32,
-  }) {
-    if (n <= 0) {
-      throw InvalidParameterException('n', n, 'n must be positive');
-    }
-    final cols = m ?? n;
-    if (cols <= 0) {
-      throw InvalidParameterException('m', cols, 'm must be positive');
-    }
-
-    final data = Float32List(n * cols);
-    final diagSize = n < cols ? n : cols;
-    for (int i = 0; i < diagSize; i++) {
-      data[i * cols + i] = 1.0;
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable([n, cols]),
-    );
-  }
+  }) =>
+      _eyeImpl(n, m: m, dtype: dtype);
 
   /// Creates a 1D tensor with evenly spaced values.
   ///
@@ -557,27 +431,8 @@ class TensorBuffer {
     double end, {
     required int steps,
     DType dtype = DType.float32,
-  }) {
-    if (steps < 1) {
-      throw InvalidParameterException('steps', steps, 'steps must be >= 1');
-    }
-
-    final data = Float32List(steps);
-
-    if (steps == 1) {
-      data[0] = start;
-    } else {
-      final step = (end - start) / (steps - 1);
-      for (int i = 0; i < steps; i++) {
-        data[i] = start + i * step;
-      }
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable([steps]),
-    );
-  }
+  }) =>
+      _linspaceImpl(start, end, steps: steps, dtype: dtype);
 
   /// Creates a 1D tensor with values in a range with a given step.
   ///
@@ -595,87 +450,26 @@ class TensorBuffer {
     required double end,
     double step = 1.0,
     DType dtype = DType.float32,
-  }) {
-    if (step == 0) {
-      throw InvalidParameterException('step', step, 'step cannot be zero');
-    }
-    if ((end > start && step < 0) || (end < start && step > 0)) {
-      throw InvalidParameterException(
-        'step',
-        step,
-        'step direction does not match range direction',
-      );
-    }
-
-    final numSteps = ((end - start) / step).ceil();
-    if (numSteps <= 0) {
-      return TensorBuffer(
-        storage: TensorStorage(Float32List(0), dtype),
-        shape: List.unmodifiable([0]),
-      );
-    }
-
-    final data = Float32List(numSteps);
-    for (int i = 0; i < numSteps; i++) {
-      data[i] = start + i * step;
-    }
-
-    return TensorBuffer(
-      storage: TensorStorage(data, dtype),
-      shape: List.unmodifiable([numSteps]),
-    );
-  }
+  }) =>
+      _arangeImpl(start: start, end: end, step: step, dtype: dtype);
 
   /// Creates a tensor from an existing [Float32List] with the given [shape].
   ///
   /// Throws [ShapeMismatchException] if data length doesn't match shape.
-  static TensorBuffer fromFloat32List(Float32List data, List<int> shape) {
-    final expectedNumel = shape.fold(1, (a, b) => a * b);
-    if (data.length != expectedNumel) {
-      throw ShapeMismatchException(
-        actual: shape,
-        message: 'Data length (${data.length}) does not match shape $shape (numel: $expectedNumel)',
-      );
-    }
-    return TensorBuffer(
-      storage: TensorStorage.fromFloat32List(data),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  static TensorBuffer fromFloat32List(Float32List data, List<int> shape) =>
+      _fromFloat32ListImpl(data, shape);
 
   /// Creates a tensor from an existing [Float64List] with the given [shape].
   ///
   /// Throws [ShapeMismatchException] if data length doesn't match shape.
-  static TensorBuffer fromFloat64List(Float64List data, List<int> shape) {
-    final expectedNumel = shape.fold(1, (a, b) => a * b);
-    if (data.length != expectedNumel) {
-      throw ShapeMismatchException(
-        actual: shape,
-        message: 'Data length (${data.length}) does not match shape $shape (numel: $expectedNumel)',
-      );
-    }
-    return TensorBuffer(
-      storage: TensorStorage.fromFloat64List(data),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  static TensorBuffer fromFloat64List(Float64List data, List<int> shape) =>
+      _fromFloat64ListImpl(data, shape);
 
   /// Creates a tensor from an existing [Uint8List] with the given [shape].
   ///
   /// Throws [ShapeMismatchException] if data length doesn't match shape.
-  static TensorBuffer fromUint8List(Uint8List data, List<int> shape) {
-    final expectedNumel = shape.fold(1, (a, b) => a * b);
-    if (data.length != expectedNumel) {
-      throw ShapeMismatchException(
-        actual: shape,
-        message: 'Data length (${data.length}) does not match shape $shape (numel: $expectedNumel)',
-      );
-    }
-    return TensorBuffer(
-      storage: TensorStorage.fromUint8List(data),
-      shape: List.unmodifiable(shape),
-    );
-  }
+  static TensorBuffer fromUint8List(Uint8List data, List<int> shape) =>
+      _fromUint8ListImpl(data, shape);
 
   /// Computes strides for a tensor with the given [shape] and [format].
   static List<int> computeStrides(List<int> shape, MemoryFormat format) {
@@ -734,103 +528,5 @@ class TensorBuffer {
   String toString() {
     return 'TensorBuffer(shape: $shape, dtype: $dtype, '
         'strides: $strides, contiguous: $isContiguous)';
-  }
-}
-
-// Math function aliases for random number generation
-double _sqrt(double x) => x >= 0 ? _power(x, 0.5) : double.nan;
-double _log(double x) {
-  if (x <= 0) return double.negativeInfinity;
-  // Natural log approximation using Taylor series or built-in
-  double result = 0;
-  double term = (x - 1) / (x + 1);
-  final termSq = term * term;
-  for (int i = 1; i <= 100; i += 2) {
-    result += term / i;
-    term *= termSq;
-  }
-  return 2 * result;
-}
-
-double _power(double base, double exp) {
-  if (exp == 0.5) {
-    // Newton's method for square root
-    if (base < 0) return double.nan;
-    if (base == 0) return 0;
-    double guess = base / 2;
-    for (int i = 0; i < 20; i++) {
-      guess = (guess + base / guess) / 2;
-    }
-    return guess;
-  }
-  // For other cases, use exp(exp * ln(base))
-  return _exp(exp * _log(base));
-}
-
-double _exp(double x) {
-  double result = 1;
-  double term = 1;
-  for (int i = 1; i <= 30; i++) {
-    term *= x / i;
-    result += term;
-    if (term.abs() < 1e-15) break;
-  }
-  return result;
-}
-
-const double _pi = 3.14159265358979323846;
-
-double _cos(double x) {
-  // Normalize to [-pi, pi]
-  while (x > _pi) {
-    x -= 2 * _pi;
-  }
-  while (x < -_pi) {
-    x += 2 * _pi;
-  }
-
-  double result = 1;
-  double term = 1;
-  final xSq = x * x;
-  for (int i = 1; i <= 15; i++) {
-    term *= -xSq / ((2 * i - 1) * (2 * i));
-    result += term;
-  }
-  return result;
-}
-
-double _sin(double x) {
-  // Normalize to [-pi, pi]
-  while (x > _pi) {
-    x -= 2 * _pi;
-  }
-  while (x < -_pi) {
-    x += 2 * _pi;
-  }
-
-  double result = x;
-  double term = x;
-  final xSq = x * x;
-  for (int i = 1; i <= 15; i++) {
-    term *= -xSq / ((2 * i) * (2 * i + 1));
-    result += term;
-  }
-  return result;
-}
-
-/// Simple seeded random number generator (Linear Congruential Generator).
-class _SeededRandom {
-  int _state;
-
-  _SeededRandom(int seed) : _state = seed & 0xFFFFFFFF;
-
-  factory _SeededRandom.system() {
-    return _SeededRandom(DateTime.now().microsecondsSinceEpoch);
-  }
-
-  double nextDouble() {
-    // LCG parameters (same as glibc)
-    _state = ((_state * 1103515245) + 12345) & 0x7FFFFFFF;
-    return _state / 0x7FFFFFFF;
   }
 }
